pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tripy"
version = "0.1.0"
description = "Enterprise-ready travel assistant backend based on FastAPI and LangGraph"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Tripy Contributors" }
]
keywords = ["fastapi", "langgraph", "ai-agent", "travel-assistant", "llm"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Framework :: FastAPI",
]
dependencies = [
  "alembic>=1.13.2",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy>=2.0.35",
  "python-jose[cryptography]>=3.3.0",
  "passlib>=1.7.4",
  "langgraph>=0.2.35",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-community>=0.3.0",
  "numpy>=1.26.0",
  "opentelemetry-api>=1.28.1",
  "opentelemetry-sdk>=1.28.1",
  "opentelemetry-exporter-otlp>=1.28.1",
  "opentelemetry-instrumentation-fastapi>=0.49b1",
  "opentelemetry-instrumentation-sqlalchemy>=0.49b1",
  "opentelemetry-instrumentation-logging>=0.49b1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.6.8",
  "mypy>=1.11.0",
  "types-python-jose>=3.3.4.20240106",
  "pre-commit>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/OiPunk/tripy"
Repository = "https://github.com/OiPunk/tripy"
Documentation = "https://github.com/OiPunk/tripy/tree/main/docs"
Issues = "https://github.com/OiPunk/tripy/issues"

[project.scripts]
tripy = "tripy.main:run"
ctrip-assistant = "ctrip_assistant.main:run"
tripy-db-upgrade = "tripy.db.migrations:upgrade_head"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
  ".git",
  ".venv",
  "graph_chat",
  "tools/__pycache__",
  "api",
  "db",
  "utils",
  "config",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
ignore_missing_imports = true
exclude = "(graph_chat|tools|api|db|utils|config|src/ctrip_assistant)"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
